worker-service/requirements.txt
requests==2.31.0
Pillow
flask==3.0.3
gunicorn==22.0.0
psutil==5.9.8
google-cloud-storage==2.16.0
google-cloud-tasks==2.16.3
